fix: geotiff export returns canvas, bbox and projection code; feat: Map coordinates interaction now supports pointer move along with click
