+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@bingo_soft/mapmanager",
-  "version": "1.58.0",
+  "version": "1.59.0",
   "description": "Map manager for GIS projects",
   "main": "src/MapManager.ts",
   "repository": {
