+++ src/Domain/Model/Interaction/InteractionCallbackType.ts
@@ -6,4 +6,4 @@ export type SelectCallbackFunction = (features: FeatureCollection) => void;
 export type ModifyCallbackFunction = (features: FeatureCollection) => void;
 export type TransformCallbackFunction = (features: FeatureCollection) => void;
 export type MeasureCallbackFunction = (measureResult: string) => void;
-export type MapCoordinatesCallbackFunction = (coordinates: number[]) => void;
+export type MapCoordinatesCallbackFunction = (coordinates: number[], mapProjection: string) => void;
